chore: change price data type from Numeric to BIGINT for bigger number support

+++ migrations/000003_create_items_table.up.sql
@@ -4,6 +4,5 @@ CREATE TABLE IF NOT EXISTS items (
     type       TEXT NOT NULL CHECK (type IN ('common', 'rare', 'legendary')),
     owner_id   TEXT NOT NULL REFERENCES guilds(id),
     available  BOOLEAN NOT NULL DEFAULT true,
-    base_price NUMERIC(20,2) NOT NULL DEFAULT 0
+    base_price BIGINT NOT NULL DEFAULT 0
 );
-
